Re-factored into modules, re-run, tests passing
