+++ src/dart_pipeline/process.py
@@ -40,7 +40,9 @@ from .util import \
     source_path, days_in_year, output_path, get_country_name, get_shapefile
 from .types import ProcessResult, PartialDate, AdminLevel
 from .constants import TERRACLIMATE_METRICS, OUTPUT_COLUMNS, MIN_FLOAT
+from .geospatial.aphroditetemperature import process_gadm_aphroditetemperature
 from .geospatial.worldpop_count import process_gadm_worldpopcount
+from .meteorological.aphroditetemperature import process_aphroditetemperature
 from .sociodemographic.worldpop_count import process_worldpopcount
 
 pandarallel.initialize(verbose=0)
@@ -228,176 +230,6 @@ def process_dengueperu(
     return master, 'dengue_peru.csv'
 
 
-def process_gadm_aphroditetemperature(
-    iso3: str, admin_level: Literal['0', '1', '2', '3'], partial_date: str,
-    resolution=['025deg'], plots=False
-):
-    """
-    Process GADM and APHRODITE Daily mean temperature product (V1808) data.
-
-    Aggregates by given admin level for the given country (ISO3 code) and
-    partial date.
-    """
-    sub_pipeline = 'geospatial/aphrodite-daily-mean-temp'
-    pdate = PartialDate.from_string(partial_date)
-    logging.info('iso3:%s', iso3)
-    logging.info('admin_level:%s', admin_level)
-    logging.info('partial_date:%s', pdate)
-    logging.info('scope:%s', pdate.scope)
-    logging.info('plots:%s', plots)
-
-    # Import shape file
-    path = get_shapefile(iso3, admin_level)
-    logging.info('importing:%s', path)
-    gdf = gpd.read_file(path)
-
-    # Initialise output data frame
-    output = pd.DataFrame(columns=OUTPUT_COLUMNS)
-
-    version = 'V1808'
-    year = pdate.year
-    params = {
-        # Parameters APHRO_MA_TAVE_025deg_V1808.ctl
-        '025deg': {
-            'product': 'TAVE',
-            'resolution': '025deg',
-            'extension': '',
-            'n_deg': (360, 280),
-            'start_coords': (60.125, -14.875),
-            'scale_factor': 0.25
-        },
-        # Parameters APHRO_MA_TAVE_050deg_V1808.nc.ctl
-        '050deg_nc': {
-            'product': 'TAVE',
-            'resolution': '050deg',
-            'extension': '.nc',
-            'n_deg': (180, 140),
-            'start_coords': (60.25, -14.75),
-            'scale_factor': 0.5
-        },
-        # Parameters APHRO_MA_TAVE_CLM_005deg_V1808.ctl
-        '005deg_nc': {
-            'product': 'TAVE_CLM',
-            'resolution': '005deg',
-            'extension': '.nc',
-            'n_deg': (1800, 1400),
-            'start_coords': (60.025, -14.975),
-            'scale_factor': 0.05
-        },
-    }
-    for data_type in resolution:
-        nday = days_in_year(int(year))
-        # Record length
-        nx, ny = params[data_type]['n_deg']
-        recl = nx * ny
-        # Longitude and latitude bounds
-        x_start, y_start = params[data_type]['start_coords']
-        scale_factor = params[data_type]['scale_factor']
-        xlon = x_start + np.arange(nx) * scale_factor
-        ylat = y_start + np.arange(ny) * scale_factor
-
-        # Open the file
-        product = params[data_type]['product']
-        res = params[data_type]['resolution']
-        ext = params[data_type]['extension']
-        path = source_path('meteorological/aphrodite-daily-mean-temp', '')
-        path = path / f'APHRO_MA_{product}_{res}_{version}.{year}{ext}'
-        # Read binary data
-        logging.info('importing:%s', path)
-        with open(path, 'rb') as f:
-            # Initialise arrays
-            temp_data = np.zeros((nday, ny, nx))
-            rstn_data = np.zeros((nday, ny, nx))
-
-            for iday in range(nday):
-                # Read next batch of temp values of size nx * ny
-                temp_raw = np.fromfile(f, dtype='float32', count=recl)
-                temp_raw = temp_raw.reshape((ny, nx))
-                # Read next batch of rstn values of size nx * ny
-                rstn_raw = np.fromfile(f, dtype='float32', count=recl)
-                rstn_raw = rstn_raw.reshape((ny, nx))
-                # Store in arrays
-                temp_data[iday, :, :] = temp_raw
-                rstn_data[iday, :, :] = rstn_raw
-
-        temp_data = temp_data.astype('float32')
-        rstn_data = rstn_data.astype('float32')
-        valid_xlon, valid_ylat = np.meshgrid(xlon, ylat, indexing='xy')
-
-        # Iterate through days
-        for iday in range(nday):
-            this_date = datetime(int(year), 1, 1) + timedelta(days=iday)
-            this_date = this_date.date()
-            # Skip if the day doesn't match the partial date
-            if pdate.month and pdate.month != this_date.month:
-                continue
-            if pdate.day and pdate.day != this_date.day:
-                continue
-
-            valid_mask = (rstn_data[iday, :, :] != 0.0) & \
-                (temp_data[iday, :, :] != NO_DATA)
-            valid_prcp = temp_data[iday][valid_mask]
-            valid_lon = valid_xlon[valid_mask]
-            valid_lat = valid_ylat[valid_mask]
-
-            # Create rows in output for each sub-region
-            to_append = []
-            for _, row in gdf.iterrows():
-                # Extract the geometry of the current sub-region (polygon)
-                region_geom = row.geometry
-
-                # Filter to get data that falls within the sub-region geometry
-                points = [
-                    shapely.geometry.Point(lon, lat) for lon, lat in
-                    zip(valid_lon, valid_lat)
-                ]
-                region_mask = np.array(
-                    [region_geom.contains(point) for point in points]
-                )
-
-                # Filter data for this sub-region
-                valid_temp_region = valid_prcp[region_mask]
-
-                output_row = {
-                    'iso3': iso3,
-                    'admin_level_0': row['COUNTRY'],
-                    'admin_level_1': row.get('NAME_1', ''),
-                    'admin_level_2': row.get('NAME_2', ''),
-                    'admin_level_3': row.get('NAME_3', ''),
-                    'year': year,
-                    'month': this_date.month,
-                    'day': this_date.day,
-                    'week': '',
-                    'value': valid_temp_region.mean() if
-                    len(valid_temp_region) > 0 else '',
-                    'resolution': '0.25°' if res == '025deg' else '0.5°',
-                    'metric': 'aphrodite-daily-mean-temp',
-                    'unit': '°C',
-                    'creation_date': date.today()
-                }
-                to_append.append(pd.DataFrame([output_row]))
-            # Concatenate the new rows to the output DataFrame
-            if to_append:
-                # Drop all-NA columns
-                to_append = [df.dropna(axis=1, how='all') for df in to_append]
-                # Drop empty data frames
-                to_append = [df for df in to_append if not df.empty]
-                output = pd.concat([output] + to_append, ignore_index=True)
-
-            # Scatter plot
-            if plots:
-                title = f'Temperature\n{this_date}'
-                colourbar_label = 'Temperature [°C]'
-                folder = f'admin_level_{admin_level}/{res.replace('0', '0_')}'
-                path = output_path(sub_pipeline) / folder / f'{this_date}.png'
-                plot_gadm_scatter(
-                    valid_lon, valid_lat, valid_prcp,
-                    title, colourbar_label, path, gdf
-                )
-
-    return output, 'aphrodite-daily-mean-temp.csv'
-
-
 def process_gadm_aphroditeprecipitation(
     iso3: str, admin_level: Literal['0', '1', '2', '3'], partial_date: str,
     resolution=['025deg', '050deg'], plots=False
@@ -576,140 +408,6 @@ def process_gadm_admin_map_data(iso3: str, admin_level: AdminLevel):
         output = pd.concat([output, new_row_df], ignore_index=True)
 
     return output, f"{iso3}/admin{admin_level}_area.csv"
-
-
-def process_aphrodite_temperature_data(year=None, plots=False) -> \
-        list[ProcessResult]:
-    """Process APHRODITE Daily mean temperature product (V1808) data."""
-    sub_pipeline = 'meteorological/aphrodite-daily-mean-temp'
-    version = 'V1808'
-    if not year:
-        # Regex pattern to match the resolution, version and year in filenames
-        pattern = r'APHRO_MA_TAVE_(\d+deg)_V(\d+)\.(\d+)'
-        # Find the latest year for which there is data
-        years = []
-        path = source_path(sub_pipeline, '')
-        for filename in Path(path).iterdir():
-            match = re.match(pattern, str(filename.name))
-            if match:
-                _, _, year = match.groups()
-                years.append(int(year))
-        # Get the latest year
-        year = str(max(years))
-
-    # Initialise output data frame
-    output = pd.DataFrame(columns=OUTPUT_COLUMNS)
-
-    params = {
-        # Parameters APHRO_MA_TAVE_025deg_V1808.ctl
-        '025deg': {
-            'product': 'TAVE',
-            'resolution': '025deg',
-            'extension': '',
-            'n_deg': (360, 280),
-            'start_coords': (60.125, -14.875),
-            'scale_factor': 0.25
-        },
-        # Parameters APHRO_MA_TAVE_050deg_V1808.nc.ctl
-        '050deg_nc': {
-            'product': 'TAVE',
-            'resolution': '050deg',
-            'extension': '.nc',
-            'n_deg': (180, 140),
-            'start_coords': (60.25, -14.75),
-            'scale_factor': 0.5
-        },
-        # Parameters APHRO_MA_TAVE_CLM_005deg_V1808.ctl
-        '005deg_nc': {
-            'product': 'TAVE_CLM',
-            'resolution': '005deg',
-            'extension': '.nc',
-            'n_deg': (1800, 1400),
-            'start_coords': (60.025, -14.975),
-            'scale_factor': 0.05
-        },
-    }
-    for data_type in ['025deg']:
-        nday = days_in_year(int(year))
-        # Record length
-        nx, ny = params[data_type]['n_deg']
-        recl = nx * ny
-        # Longitude and latitude bounds
-        x_start, y_start = params[data_type]['start_coords']
-        scale_factor = params[data_type]['scale_factor']
-        xlon = x_start + np.arange(nx) * scale_factor
-        ylat = y_start + np.arange(ny) * scale_factor
-
-        # Open the file
-        product = params[data_type]['product']
-        res = params[data_type]['resolution']
-        ext = params[data_type]['extension']
-        path = source_path(sub_pipeline, '')
-        path = path / f'APHRO_MA_{product}_{res}_{version}.{year}{ext}'
-        # Read binary data
-        logging.info('opening:%s', path)
-        with open(path, 'rb') as f:
-            # Initialise arrays
-            temp_data = np.zeros((nday, ny, nx))
-            rstn_data = np.zeros((nday, ny, nx))
-
-            for iday in range(nday):
-                # Read next batch of temp values of size nx * ny
-                temp_raw = np.fromfile(f, dtype='float32', count=recl)
-                temp_raw = temp_raw.reshape((ny, nx))
-                # Read next batch of rstn values of size nx * ny
-                rstn_raw = np.fromfile(f, dtype='float32', count=recl)
-                rstn_raw = rstn_raw.reshape((ny, nx))
-                # Store in arrays
-                temp_data[iday, :, :] = temp_raw
-                rstn_data[iday, :, :] = rstn_raw
-
-        temp_data = temp_data.astype('float32')
-        rstn_data = rstn_data.astype('float32')
-        valid_xlon, valid_ylat = np.meshgrid(xlon, ylat, indexing='xy')
-
-        # Iterate through days
-        for iday in range(nday):
-            this_date = datetime(int(year), 1, 1) + timedelta(days=iday)
-            this_date = this_date.date()
-            valid_mask = (rstn_data[iday, :, :] != 0.0) & \
-                (temp_data[iday, :, :] != NO_DATA)
-            valid_temp = temp_data[iday][valid_mask]
-            valid_lon = valid_xlon[valid_mask]
-            valid_lat = valid_ylat[valid_mask]
-
-            # Scatter plot
-            if plots:
-                title = f'Temperature\n{this_date}'
-                colourbar_label = 'Temperature [°C]'
-                folder = res.replace('0', '0_')
-                path = output_path(sub_pipeline) / folder / f'{this_date}.png'
-                plot_scatter(
-                    valid_lon, valid_lat, valid_temp, title, colourbar_label,
-                    path
-                )
-
-            i = len(output)
-            output.loc[i, 'year'] = year
-            output.loc[i, 'month'] = this_date.month
-            output.loc[i, 'day'] = this_date.day
-            output.loc[i, 'value'] = valid_temp.mean()
-            if res == '025deg':
-                output.loc[i, 'resolution'] = '0.25°'
-            elif res == '050deg':
-                output.loc[i, 'resolution'] = '0.5°'
-
-    output['iso3'] = ''
-    output['admin_level_0'] = ''
-    output['admin_level_1'] = ''
-    output['admin_level_2'] = ''
-    output['admin_level_3'] = ''
-    output['week'] = ''
-    output['metric'] = 'aphrodite-daily-mean-temp'
-    output['unit'] = '°C'
-    output['creation_date'] = date.today()
-
-    return output, 'aphrodite-daily-mean-temp.csv'
 
 
 def process_aphrodite_precipitation_data(
@@ -1250,7 +948,7 @@ PROCESSORS: dict[str, Callable[..., ProcessResult | list[ProcessResult]]] = {
     'geospatial/chirps-rainfall': process_gadm_chirps_rainfall,
     'geospatial/gadm': process_gadm_admin_map_data,
     'geospatial/worldpop-count': process_gadm_worldpopcount,
-    'meteorological/aphrodite-daily-mean-temp': process_aphrodite_temperature_data,
+    'meteorological/aphrodite-daily-mean-temp': process_aphroditetemperature,
     'meteorological/aphrodite-daily-precip': process_aphrodite_precipitation_data,
     'meteorological/chirps-rainfall': process_chirps_rainfall,
     'meteorological/era5-reanalysis': process_era5_reanalysis_data,
