+++ tests/test_process.py
@@ -12,10 +12,8 @@ import pytest
 from dart_pipeline.process import \
     process_rwi, \
     process_dengueperu, \
-    process_gadm_aphroditetemperature, \
     process_gadm_aphroditeprecipitation, \
     process_gadm_chirps_rainfall, \
-    process_aphrodite_temperature_data, \
     process_aphrodite_precipitation_data, \
     process_chirps_rainfall, \
     process_terraclimate
@@ -161,74 +159,6 @@ class MockFile(BytesIO):
         return 1
 
 
-def test_process_gadm_aphroditetemperature():
-    iso3 = 'VNM'
-    admin_level = '0'
-    partial_date = '2023-07'
-    resolution = ['025deg']
-    plots = False
-
-    with patch('dart_pipeline.process.PartialDate') as mock_partial_date, \
-         patch('dart_pipeline.util.get_shapefile') as mock_get_shapefile, \
-         patch('geopandas.read_file') as mock_read_file, \
-         patch('dart_pipeline.util.source_path') as mock_source_path, \
-         patch('dart_pipeline.util.output_path') as mock_output_path, \
-         patch('builtins.open') as mock_open, \
-         patch('numpy.fromfile') as mock_fromfile:
-
-        # Mock PartialDate to return a specific year when accessed
-        mock_partial_date.from_string.return_value = MagicMock(
-            year=2023, month=7, day=None, scope='year'
-        )
-
-        # Mock shapefile loading
-        mock_get_shapefile.return_value = 'mock_shapefile_path'
-
-        # Create a mock GeoDataFrame row with a geometry attribute
-        mock_row = MagicMock()
-        mock_row.geometry = Polygon([(0, 0), (1, 0), (1, 1), (0, 1)])
-        mock_row.COUNTRY = 'Vietnam'
-        mock_row.NAME_1 = 'Mock Province'
-        mock_row.NAME_2 = 'Mock District'
-        mock_row.NAME_3 = 'Mock Sub-district'
-
-        # Mock geopandas dataframe and its iterrows method
-        mock_gdf = MagicMock()
-        mock_gdf.iterrows.return_value = iter([(0, mock_row)])
-        mock_read_file.return_value = mock_gdf
-
-        # Mock source_path and output_path
-        mock_source_path.return_value = MagicMock()
-        mock_output_path.return_value = MagicMock()
-
-        # Mock np.fromfile() to return a fake array
-        nx, ny = 360, 280
-        recl = nx * ny
-
-        # Create a fake array with the correct number of values
-        fake_array = np.ones(recl, dtype='float32')
-
-        # Mock np.fromfile() to return the fake array when called
-        mock_fromfile.return_value = fake_array
-
-        # Create a mock file object
-        mock_file = MockFile()
-        mock_open.return_value.__enter__.return_value = mock_file
-
-        # Call the function
-        output, csv_path = process_gadm_aphroditetemperature(
-            iso3, admin_level, partial_date, resolution, plots
-        )
-
-        # Assertions
-        assert isinstance(output, pd.DataFrame)
-        assert 'iso3' in output.columns
-        assert 'value' in output.columns
-        assert output['iso3'].iloc[0] == iso3
-        assert output['value'].iloc[0] == ''
-        assert csv_path == 'aphrodite-daily-mean-temp.csv'
-
-
 def test_process_gadm_aphroditeprecipitation():
     iso3 = 'VNM'
     admin_level = '0'
@@ -372,45 +302,6 @@ def test_process_gadm_chirps_rainfall(
 
     # Verify the correct handling of bounds and plot creation
     min_lon, min_lat, max_lon, max_lat = region_geometry.bounds
-
-
-def test_process_aphrodite_temperature_data():
-    # Minimal mocking for `np.fromfile` and file operations
-    nx, ny, _ = 360, 280, 365
-    # Mock temperature data
-    mock_temp = np.full((ny, nx), 25.0, dtype='float32')
-    # Mock station count data
-    mock_rstn = np.ones((ny, nx), dtype='float32')
-
-    def mock_fromfile(file, dtype, count):
-        if dtype == 'float32' and count == nx * ny:
-            return mock_temp.flatten() \
-                if 'temp' in file.name else mock_rstn.flatten()
-        raise ValueError(
-            f'Unexpected call to np.fromfile with {file}, {dtype}, {count}'
-        )
-
-    # Mock file opening
-    mocked_open = mock_open()
-    with patch('builtins.open', mocked_open), \
-            patch('numpy.fromfile', mock_fromfile):
-        # Call the function
-        year = 2023
-        output, csv_name = process_aphrodite_temperature_data(
-            year=year, plots=False
-        )
-
-        # Assert the output is a DataFrame
-        assert isinstance(output, pd.DataFrame)
-        assert len(output) > 0  # Ensure some data is processed
-        assert 'year' in output.columns
-        assert 'value' in output.columns
-
-        # Check key output values
-        assert (output['year'] == year).all()
-        assert (output['metric'] == 'aphrodite-daily-mean-temp').all()
-        assert (output['unit'] == '°C').all()
-        assert csv_name == 'aphrodite-daily-mean-temp.csv'
 
 
 def test_process_aphrodite_precipitation_data():
